requirements.txt
numpy
joblib
scikit-learn
pandas
psycopg2-binary
kafka-python
pyspark